refactor: extract map viewer logic
