+++ src/index.css
@@ -27,3 +27,136 @@ a {
 .scrollbar-stable {
   scrollbar-gutter: stable;
 }
+@import 'leaflet/dist/leaflet.css';
+
+/* Add custom styles for the map */
+#map {
+  height: 100%;
+  width: 100%;
+  z-index: 0;
+}
+
+/* Ensure Leaflet tiles display correctly */
+.leaflet-tile-pane {
+  z-index: 0;
+}
+
+/* Keep markers below overlay components like legends */
+.leaflet-marker-pane {
+  z-index: 5; /* keep markers behind overlay elements */
+}
+
+.leaflet-control-container {
+  z-index: 10;
+}
+
+/* Display popups above overlay components */
+.leaflet-popup-pane {
+  z-index: 1000; /* high value so popups show over legends */
+}
+
+/* Style for the water bottle icons */
+.water-bottle-icon {
+  filter: drop-shadow(0 1px 3px rgba(0, 0, 0, 0.4));
+}
+
+/* Style the popup to match the app design */
+/* Dark mode popup styles will be applied via JS in updatePopupStyles */
+.leaflet-popup-content-wrapper {
+  border-radius: 8px;
+  box-shadow: 0 3px 14px rgba(0, 0, 0, 0.2);
+  padding: 0;
+}
+
+.leaflet-popup-content {
+  margin: 18px 20px;
+  font-family: inherit;
+  line-height: 1.5;
+  width: 280px;
+  max-width: 320px;
+}
+
+/* Fix close button position */
+.leaflet-popup-close-button {
+  top: 8px !important;
+  right: 8px !important;
+  color: inherit !important;
+  opacity: 0.7;
+  transition: opacity 0.2s ease;
+  width: 18px !important;
+  height: 18px !important;
+  padding: 0 !important;
+  display: flex !important;
+  align-items: center;
+  justify-content: center;
+}
+
+.leaflet-popup-close-button:hover {
+  opacity: 1;
+  background: transparent !important;
+}
+
+/* Site popup styling */
+.site-popup {
+  padding: 4px 2px;
+}
+
+.site-popup .site-name {
+  word-wrap: break-word;
+  max-width: 100%;
+}
+
+/* Quality indicators */
+.site-popup .text-green-500 {
+  color: #22c55e;
+}
+
+.site-popup .text-yellow-400 {
+  color: #facc15;
+}
+
+.site-popup .text-red-500 {
+  color: #ef4444;
+}
+
+/* Light mode styles (default) */
+.leaflet-popup-content-wrapper {
+  background-color: rgba(255, 255, 255, 0.95);
+  color: #1f2937;
+}
+
+.site-popup .opacity-75 {
+  color: #6b7280; /* text-gray-500 */
+}
+
+/* Ensure consistent spacing between popup elements */
+.site-popup .mt-1 {
+  margin-top: 0.375rem;
+}
+
+.site-popup .mt-2 {
+  margin-top: 0.75rem;
+}
+
+/* Dark mode styles (applied via JS) */
+.dark-mode-popup .leaflet-popup-content-wrapper {
+  background-color: rgb(31, 41, 55);
+  color: rgb(209, 213, 219);
+}
+
+.dark-mode-popup .leaflet-popup-tip {
+  background-color: rgb(31, 41, 55);
+}
+
+.dark-mode-popup .site-popup .opacity-75 {
+  color: #9ca3af; /* text-gray-400 */
+}
+
+.dark-mode-popup .leaflet-popup-close-button {
+  color: rgb(209, 213, 219) !important;
+}
+
+/* Hide or minimize any remaining attribution text */
+.leaflet-control-attribution {
+  display: none;
+}
